Добавил категории

+++ DocumentDirectoryWeb/Models/DocumentCategory.cs
@@ -4,7 +4,7 @@ namespace DocumentDirectoryWeb.Models;
 
 public class DocumentCategory
 {
-    [Key] public long Id { get; set; }
+    [Key] public int Id { get; set; }
 
     [Required(ErrorMessage = "Поле является обязательным.")]
     [Display(Name = "Название")]
